BKRS-365-rename-package - renamed secret agent package
