+++ config_secret_agent_test.go
@@ -17,7 +17,7 @@ package backup
 import (
 	"testing"
 
-	saClient "github.com/aerospike/backup-go/pkg/secret-agent"
+	saClient "github.com/aerospike/backup-go/pkg/secretagent"
 	"github.com/stretchr/testify/require"
 )
