+++ pkg/secretagent/connection/connection.go
@@ -23,7 +23,7 @@ import (
 	"net"
 	"time"
 
-	"github.com/aerospike/backup-go/pkg/secret-agent/models"
+	"github.com/aerospike/backup-go/pkg/secretagent/models"
 )
 
 // magic const is taken from the Secret Agent service. It is used
