+++ secret_agent.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 	"time"
 
-	saClient "github.com/aerospike/backup-go/pkg/secret-agent"
+	saClient "github.com/aerospike/backup-go/pkg/secretagent"
 )
 
 const secretPrefix = "secrets:"
